pyproject.toml
[project]
name = "blacklist-my"
version = "0.1.0"
description = ""
authors = [
    {name = "a.dolgov",email = "dolgov_am@mail.ru"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pytelegrambotapi (>=4.27.0,<5.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "aiohttp (>=3.12.2,<4.0.0)",
    "utils (>=1.0.2,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
